+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [Unreleased]
+### Changed
+* Created new Statement interface including both old Stmt and new statementBuilder struct for building queries
+* Added new StatementQueryParameters struct for simple and safe handling of query parameters
 
 ## [0.11.1] - 2023-03-01
 ### Changed